+++ src/julia.c
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 14:16:40 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/10/10 14:49:46 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/10/18 19:03:05 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 int	julia(int w, int h, t_vars *vars)
 {
 	int		count;
-	double	c[2];
-	double	*z;
-	double	tmp;
+	long double	c[2];
+	long double	*z;
+	long double	tmp;
 
 	to_z(w, h, c, vars);
 	count = 0;
